release 3.1.200 source code

+++ services/modelarts/src/main/java/com/huaweicloud/sdk/modelarts/v1/model/CreateWorkspaceResponse.java
@@ -32,12 +32,12 @@ public class CreateWorkspaceResponse extends SdkResponse {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "update_time")
 
-    private Integer updateTime;
+    private Long updateTime;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "create_time")
 
-    private Integer createTime;
+    private Long createTime;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "enterprise_project_name")
@@ -125,7 +125,7 @@ public class CreateWorkspaceResponse extends SdkResponse {
         this.enterpriseProjectId = enterpriseProjectId;
     }
 
-    public CreateWorkspaceResponse withUpdateTime(Integer updateTime) {
+    public CreateWorkspaceResponse withUpdateTime(Long updateTime) {
         this.updateTime = updateTime;
         return this;
     }
@@ -134,15 +134,15 @@ public class CreateWorkspaceResponse extends SdkResponse {
      * 最后修改时间，UTC。
      * @return updateTime
      */
-    public Integer getUpdateTime() {
+    public Long getUpdateTime() {
         return updateTime;
     }
 
-    public void setUpdateTime(Integer updateTime) {
+    public void setUpdateTime(Long updateTime) {
         this.updateTime = updateTime;
     }
 
-    public CreateWorkspaceResponse withCreateTime(Integer createTime) {
+    public CreateWorkspaceResponse withCreateTime(Long createTime) {
         this.createTime = createTime;
         return this;
     }
@@ -151,11 +151,11 @@ public class CreateWorkspaceResponse extends SdkResponse {
      * 创建时间，UTC。
      * @return createTime
      */
-    public Integer getCreateTime() {
+    public Long getCreateTime() {
         return createTime;
     }
 
-    public void setCreateTime(Integer createTime) {
+    public void setCreateTime(Long createTime) {
         this.createTime = createTime;
     }
